requirements.txt
PyVCF3
pandas
requests
matplotlib
seaborn
plotly
dash
dash-bootstrap-components
scikit-learn
joblib
pytest
pyarrow 